Split director package
To keep this manageable, split the director package out into a handful
of modules.
